+++ CHANGELOG.md
@@ -1,9 +1,20 @@
+## 0.1.0 - 24/03/2022
+## 0.0.9 - 23/03/2022
+## 0.0.8 - 20/03/2022
+## 0.0.7 - 20/03/2022
+## 0.0.6 - 20/03/2022
+## 0.0.5 - 17/03/2022
+## 0.0.4 - 17/03/2022
+## 0.0.3 - 17/03/2022
+## 0.0.2 - 13/03/2022
 ## 0.0.1 - 12/03/2022
 ## 0.0.2 - 13/03/2022
 
 
 * TODO: My Person Code
 * TODO: Context,image,integer,string extension added, and Email Regexp and Random Color constants added.
+* TODO: Context,image,integer,string,vb... extension added, and Email Regexp and Random Color constants added.
+* TODO: Add Dynamic Text Widget.
 
 * TODO: version 0.0.2 upgraded string extension and depends components